refactor: FeignClient 동작 여부 확인 및 추가 설정 - UserFeignClient: hub로 라우팅 변경 - GatewayConfig: 라우팅 변경

+++ auth/src/main/java/com/two_ddang/logistics/auth/service/JwtTokenGenerator.java
@@ -24,7 +24,7 @@ public class JwtTokenGenerator {
         this.expirationTime = expirationTime;
     }
 
-    public String createJwtToken(Long userId, String email,
+    public String createJwtToken(Integer userId, String email,
                                  String username, UserType userType) {
         return Jwts.builder()
                 .claim("userId", userId)
